Change the way that arguments are handled
Change the way the scrolling speed and GIF width arguments are
handled. Now, there are 2 argument groups that allow you to select a
unit of measurement and a value. The avalibale options remain the
same, but the interface is much more sensible now, not requiring
manually setting unused options to -1.

For more information see, the -h otpur of the script