Added GIF playback server
Added GIF playback server, a program that listens on a ZeroMQ socket
for incoming GIFs and then plays them on the AniMe matrix via
asusctl. It maintains an internal priority queue of GIFs and sends
them to the matrix according to their priority.

Please note that the server relies on ZeroMQ and Pickle, which
in theory, give anyone with access to the socket an ability to run
arbitrary python code in the context of the server. The correct thing
to do here is to run the server with limited access to stuff with e.g.
bubblewrap.